+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.12.1] - Mar 23, 2023
+
+### Added
+
+- Add quotes page to the header and footer
+
+### Changed
+
+- Retry until a successfull response is recieved when the book page scraper encounters a 504 error
+- Remove unneeded route index pages
+
 ## [2.12.0] - Mar 13, 2023
 
 ### Added